Should be parsing floats instead of ints

+++ example/js2/app/src/index.js
@@ -49,7 +49,7 @@ class NumberTemplate extends ParameterTemplate {
 	}
 	
 	getEventValue(event) {
-		return parseInt(event.target.value);
+		return parseFloat(event.target.value);
 	}
 }
